% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{make_summary}
\alias{make_summary}
\title{Summarize data}
\usage{
make_summary(df, groupvars, stats, width, measure)
}
\arguments{
\item{df}{Data Frame}

\item{groupvars}{Name of the column(s) containing your bsfactor and wsfactor}

\item{stats}{Summary Statistic to use, see documentation for a list.}

\item{width}{Type of error bar ("SE" or "CI")}

\item{measure}{The name of the column containing your dependent variable}
}
\value{
data.frame
}
\description{
Summarizes data using summary statistic of choice and returns a data frame sorted by group.
}
\examples{
make_summary(ToothGrowth, groupvars = c("supp", "dose"), stats = "mean", width = "CI", measure = "len")
}
